[Python O11y] Remove exporter from public API and refactor tests
