+++ src/python/grpcio_observability/grpc_observability/_open_telemetry_observability.py
@@ -62,7 +62,6 @@ class OpenTelemetryObservability(grpc._observability.ObservabilityPlugin):
     This is class is part of an EXPERIMENTAL API.
 
     Args:
-      exporter: Exporter used to export data.
       plugin: OpenTelemetryPlugin to enable.
     """
 
@@ -80,10 +79,7 @@ class OpenTelemetryObservability(grpc._observability.ObservabilityPlugin):
             for plugin in plugins:
                 _plugins.append(_OpenTelemetryPlugin(plugin))
 
-        if exporter:
-            self.exporter = exporter
-        else:
-            self.exporter = _OpenTelemetryExporterDelegator(_plugins)
+        self.exporter = _OpenTelemetryExporterDelegator(_plugins)
 
         try:
             _cyobservability.activate_stats()
